# src/docs/openapi.yaml
openapi: "3.0.0"
info:
  title: Shopix E-Commerce API
  version: "1.0.0"
  description: |
    Full REST API documentation for the **Shopix** e-commerce platform.

    ### Authentication
    Most endpoints require a valid `Accesstoken` cookie set after login.
    Pass credentials via the login endpoint — the server sets HttpOnly cookies automatically.

  contact:
    name: Shopix Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local Development
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: Accesstoken
  schemas:
...
    Store:
      type: object
      properties:
        _id:
          type: string
        sellerId:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
        logo:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, frozen]
        createdAt:
          type: string
          format: date-time

    # ─── Conversation ─────────────────────────────────────────
...
paths:
...
  /api/wishlist/move-to-cart:
    post:
      tags: [Wishlist]
      summary: Move wishlist item to cart
      description: Adds the product to the cart (if not already there) and removes it from the wishlist.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, wishlistId]
              properties:
                productId:
                  type: string
                wishlistId:
                  type: string
      responses:
        "200":
          description: Moved to cart

  # ───────────────────────────────────────────────────────────
  # ORDERS
  # ───────────────────────────────────────────────────────────
...
  /api/store:
    get:
      tags: [Store]
      summary: Get seller's store
      description: Returns the authenticated seller's store details.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Store found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  store:
                    $ref: "#/components/schemas/Store"
        "404":
          description: No store found
        "401":
          description: Unauthorized

    post:
      tags: [Store]
      summary: Create or re-submit store application
      description: |
        Creates a new store application (status `pending`). If a rejected store exists, re-submits it.
        Notifies admin via email and in-app notification.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, description, type, logo]
              properties:
                name:
                  type: string
                  minLength: 4
                  example: My Gadget Store
                description:
                  type: string
                  description: Max 100 words
                type:
                  type: string
                  example: Electronics
                logo:
                  type: string
                  format: binary
      responses:
        "201":
          description: Store submitted for approval
        "200":
          description: Store re-submitted for approval
        "400":
          description: Validation error or store already exists
...
  /api/admin/dashboard:
    get:
      tags: [Admin – Dashboard]
      summary: Admin dashboard stats
      description: Returns total users, total stores, and monthly registration chart data for the last 6 months. Admin only.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      totalUsers:
                        type: integer
                      totalStores:
                        type: integer
                      chartData:
                        type: array
                        items:
                          type: object
                          properties:
                            name:
                              type: string
                              example: Jan
                            users:
                              type: integer
                            stores:
                              type: integer
                      admin:
                        type: object
                        properties:
                          name:
                            type: string
                          email:
                            type: string
        "403":
          description: Unauthorized — admin only

  # ───────────────────────────────────────────────────────────
  # ADMIN – USERS
  # ───────────────────────────────────────────────────────────
...